# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/escrows": {
      "get": {
        "operationId": "EscrowsController_findAll_v1",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "meta": {
                          "$ref": "#/components/schemas/PaginationMetaDto"
                        },
                        "result": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Escrow"
                          }
                        }
                      },
                      "required": [
                        "meta",
                        "result"
                      ]
                    }
                  },
                  "required": [
                    "statusCode",
                    "data"
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "Escrows"
        ]
      }
    },
    "/api/v1/escrows/order/{orderId}": {
      "get": {
        "operationId": "EscrowsController_findByOrder_v1",
        "parameters": [
          {
            "name": "orderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Escrow"
                      }
                    }
                  },
                  "required": [
                    "statusCode",
                    "data"
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "Escrows"
        ]
      }
    },
    "/api/v1/escrows/seller/{sellerId}": {
      "get": {
        "operationId": "EscrowsController_findBySeller_v1",
        "parameters": [
          {
            "name": "sellerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "meta": {
                          "$ref": "#/components/schemas/PaginationMetaDto"
                        },
                        "result": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Escrow"
                          }
                        }
                      },
                      "required": [
                        "meta",
                        "result"
                      ]
                    }
                  },
                  "required": [
                    "statusCode",
                    "data"
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "Escrows"
        ]
      }
    },
    "/api/v1/escrows/held/{sellerId}": {
      "get": {
        "operationId": "EscrowsController_getHeldBalance_v1",
        "parameters": [
          {
            "name": "sellerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "held_balance": {
                          "type": "number"
                        }
                      },
                      "required": [
                        "held_balance"
                      ]
                    }
                  },
                  "required": [
                    "statusCode",
                    "data"
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "Escrows"
        ]
      }
    },
...
    "/api/v1/payments": {
      "post": {
        "operationId": "PaymentsController_create_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePaymentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 201
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Payment"
                    }
                  },
                  "required": [
                    "statusCode",
                    "data"
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "Payments"
        ]
      },
      "get": {
        "operationId": "PaymentsController_findAll_v1",
        "parameters": [
          {
            "name": "currentPage",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "meta": {
                          "$ref": "#/components/schemas/PaginationMetaDto"
                        },
                        "result": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Payment"
                          }
                        }
                      },
                      "required": [
                        "meta",
                        "result"
                      ]
                    }
                  },
                  "required": [
                    "statusCode",
                    "data"
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "Payments"
        ]
      }
    },
...
    "/api/v1/payments/{id}": {
      "get": {
        "operationId": "PaymentsController_findOne_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Payment"
                    }
                  },
                  "required": [
                    "statusCode",
                    "data"
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "Payments"
        ]
      },
      "patch": {
        "operationId": "PaymentsController_update_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePaymentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Payment"
                    }
                  },
                  "required": [
                    "statusCode",
                    "data"
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "Payments"
        ]
      },
      "delete": {
        "operationId": "PaymentsController_remove_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "statusCode",
                    "data"
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "Payments"
        ]
      }
    },
...
  },
  "info": {
    "title": "Zoldify API",
    "description": "API của sàn mua bán đồ cũ Zoldify. Web (Next.js) và app (React Native) dùng chung hợp đồng này.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "PaginationMetaDto": {
        "type": "object",
        "properties": {
          "current": {
            "type": "number",
            "description": "Trang hiện tại, bắt đầu từ 1"
          },
          "pageSize": {
            "type": "number",
            "description": "Số bản ghi mỗi trang"
          },
          "pages": {
            "type": "number",
            "description": "Tổng số trang"
          },
          "total": {
            "type": "number",
            "description": "Tổng số bản ghi khớp điều kiện lọc"
          }
        },
        "required": [
          "current",
          "pageSize",
          "pages",
          "total"
        ]
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "full_name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone_number": {
            "type": "string"
          },
          "role": {
            "enum": [
              "buyer",
              "seller",
              "admin",
              "moderator"
            ],
            "type": "string"
          },
          "avatar": {
            "type": "string"
          },
          "email_verified": {
            "type": "boolean"
          },
          "is_locked": {
            "type": "boolean"
          },
          "last_seen": {
            "format": "date-time",
            "type": "string"
          },
          "gender": {
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "deleted_at": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "full_name",
          "email",
          "phone_number",
          "role",
          "avatar",
          "email_verified",
          "is_locked",
          "last_seen",
          "gender",
          "created_at"
        ]
      },
...
      "Product": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "currency": {
            "type": "string",
            "description": "4.1. Đơn vị tiền của `price` — mã ISO 4217 ('VND', 'USD', 'JPY'…).\n\nVÌ SAO CẦN: trước cột này, mọi con số trong hệ thống đều NGẦM hiểu là đồng\nViệt Nam, và không chỗ nào trong mã nói ra điều đó. Giao diện tiếng Anh vì\nvậy vẫn in \"1.890.000 ₫\" — không phải lỗi định dạng, mà là vì thật sự không\ncó dữ liệu nào cho nó biết số đó là tiền gì.\n\n⚠️ CỘT NÀY CHỈ LÀM ĐÚNG PHẦN HIỂN THỊ. Nó KHÔNG khiến sàn giao dịch được\nxuyên tiền tệ. Muốn thế còn cần: nguồn tỉ giá, thời điểm chốt giá, và một\nsổ cái nhiều tiền tệ. Sổ cái hiện tại (ledger_accounts) kiểm bất biến\n`sum(entries) = 0`, mà bất biến đó chỉ có nghĩa TRONG CÙNG một tiền tệ —\ntrộn hai loại tiền vào một sổ là cách làm hỏng sổ sách mà không ai thấy.\nNên tới khi có thiết kế đó, mỗi sàn vẫn nên chạy một tiền tệ."
          },
          "stock": {
            "type": "number"
          },
          "image": {
            "type": "string"
          },
          "brand": {
            "type": "string"
          },
          "spec": {
            "type": "string"
          },
          "images": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "condition": {
            "type": "string"
          },
          "is_freeship": {
            "type": "boolean"
          },
          "sold_count": {
            "type": "number"
          },
          "view_count": {
            "type": "number"
          },
          "status": {
            "enum": [
              "draft",
              "pending",
              "active",
              "sold",
              "rejected"
            ],
            "type": "string"
          },
          "category": {
            "$ref": "#/components/schemas/Category"
          },
          "seller": {
            "$ref": "#/components/schemas/User"
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string"
          },
          "deleted_at": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "description",
          "price",
          "currency",
          "stock",
          "image",
          "brand",
          "spec",
          "images",
          "condition",
          "is_freeship",
          "sold_count",
          "view_count",
          "status",
          "category",
          "seller",
          "created_at",
          "updated_at"
        ]
      },
...
      "Order": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "order_code": {
            "type": "string"
          },
          "idempotency_key": {
            "type": "string",
            "nullable": true,
            "description": "Khoá chống trùng — một giỏ hàng chỉ đặt được đúng một đơn.\n\nSinh từ `sha256(userId + ':' + các id dòng giỏ hàng đã sắp xếp)`. Người mua\nbấm \"Đặt hàng\" hai lần cùng lúc thì cả hai cùng sinh ra chuỗi này, và khoá\nUNIQUE dưới database để đúng một lượt đi qua. Lượt kia nhận lại chính đơn\nvừa tạo, không phải một thông báo lỗi.\n\nĐo được trước khi có nó, bằng `npm run check:race` R5: 20 lượt bấm đồng\nthời → 20 đơn, kho trừ 20 lần.\n\nNULL cho đơn có trước migration `1787900000000`. MySQL cho phép nhiều NULL\ntrong một khoá UNIQUE nên chúng không đụng nhau."
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "total_amount": {
            "type": "number"
          },
          "shipping_fee": {
            "type": "number"
          },
          "discount_amount": {
            "type": "number"
          },
          "final_amount": {
            "type": "number"
          },
          "currency": {
            "type": "string",
            "description": "Đơn vị tiền của đơn — CHỤP LẠI lúc đặt, không đọc lại từ sản phẩm.\n\nNgười bán đổi tiền tệ của tin đăng sau khi đơn đã đặt là chuyện có thật, và\nkhi đó đơn cũ phải giữ nguyên thứ nó đã thoả thuận. Đọc `product.currency`\nlúc hiển thị thì một đơn 1.890.000 VND có thể biến thành 1.890.000 USD chỉ\nvì người bán sửa tin — hoá đơn tự đổi số sau lưng người mua.\n\nCùng lý do mà `order_items` đã chụp lại tên và giá sản phẩm."
          },
          "status": {
            "enum": [
              "pending",
              "confirmed",
              "processing",
              "shipping",
              "delivered",
              "cancelled",
              "refunded"
            ],
            "type": "string"
          },
          "payment_method": {
            "enum": [
              "cod",
              "bank_transfer",
              "wallet",
              "momo",
              "vnpay",
              "payos"
            ],
            "type": "string"
          },
          "is_paid": {
            "type": "boolean"
          },
          "paid_at": {
            "format": "date-time",
            "type": "string"
          },
          "receiver_name": {
            "type": "string"
          },
          "receiver_phone": {
            "type": "string"
          },
          "shipping_address": {
            "type": "string"
          },
          "province": {
            "type": "string"
          },
          "district": {
            "type": "string"
          },
          "ghn_district_id": {
            "type": "number"
          },
          "ghn_ward_code": {
            "type": "string"
          },
          "note": {
            "type": "string"
          },
          "tracking_code": {
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string"
          },
          "deleted_at": {
            "format": "date-time",
            "type": "string"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItem"
            }
          }
        },
        "required": [
          "id",
          "order_code",
          "idempotency_key",
          "user",
          "total_amount",
          "shipping_fee",
          "discount_amount",
          "final_amount",
          "currency",
          "status",
          "payment_method",
          "is_paid",
          "paid_at",
          "receiver_name",
          "receiver_phone",
          "shipping_address",
          "province",
          "district",
          "ghn_district_id",
          "ghn_ward_code",
          "note",
          "tracking_code",
          "created_at",
          "updated_at",
          "items"
        ]
      },
      "OrderItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "order": {
            "$ref": "#/components/schemas/Order"
          },
          "product": {
            "$ref": "#/components/schemas/Product"
          },
          "product_name": {
            "type": "string"
          },
          "product_image": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "quantity": {
            "type": "number"
          },
          "subtotal": {
            "type": "number"
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "order",
          "product",
          "product_name",
          "product_image",
          "price",
          "quantity",
          "subtotal",
          "created_at",
          "updated_at"
        ]
      },
...
      "Escrow": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "order": {
            "$ref": "#/components/schemas/Order"
          },
          "buyer": {
            "$ref": "#/components/schemas/User"
          },
          "seller": {
            "$ref": "#/components/schemas/User"
          },
          "amount": {
            "type": "number",
            "description": "Tiền HÀNG của người bán này — tổng `subtotal` các món của họ. Phí sàn tính trên đúng số này."
          },
          "shipping_amount": {
            "type": "number",
            "description": "Phần PHÍ SHIP của người bán này, tách riêng khỏi tiền hàng.\n\nVÌ SAO PHẢI CÓ CỘT RIÊNG. Người mua trả `final_amount` (hàng + ship) và cả\nsố đó chảy vào `platform/escrow_hold`. Nhưng trước 24/09 khoản ký quỹ chỉ\nghi tiền hàng, nên phần ship KHÔNG CÓ ĐƯỜNG NÀO RA khỏi két: giải ngân thì\nngười bán không nhận được nó, hoàn tiền thì người mua cũng không. Nó đọng\nlại mãi mãi, mỗi đơn một ít. Đo bằng TC-P0-04a/b.\n\nVÌ SAO KHÔNG GỘP VÀO `amount`. Vì phí sàn tính trên `amount`, mà theo quyết\nđịnh của Đạt (24/09): *\"Tiền ship bên mua bán họ tự trả chứ sàn không thu\ntiền\"*. Sàn chỉ cầm hộ rồi chuyển đi — không được ăn phần trăm trên đó.\nGộp hai con số vào một cột là mất đúng ranh giới ấy.\n\nĐơn cũ (trước migration) để 0, nên tiền ship của chúng vẫn kẹt — cần đối\nsoát tay nếu production đã có đơn thật."
          },
          "status": {
            "enum": [
              "holding",
              "released",
              "refunded",
              "cancelled"
            ],
            "type": "string"
          },
          "released_at": {
            "format": "date-time",
            "type": "string"
          },
          "note": {
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "order",
          "buyer",
          "seller",
          "amount",
          "shipping_amount",
          "status",
          "released_at",
          "note",
          "created_at",
          "updated_at"
        ]
      },
...
      "CreatePaymentDto": {
        "type": "object",
        "properties": {
          "order_id": {
            "type": "number"
          },
          "amount": {
            "type": "number",
            "minimum": 1000
          },
          "payment_method": {
            "type": "string",
            "enum": [
              "cod",
              "bank_transfer",
              "wallet",
              "momo",
              "vnpay",
              "payos"
            ]
          }
        }
      },
      "Payment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "order": {
            "$ref": "#/components/schemas/Order"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "amount": {
            "type": "number"
          },
          "payment_method": {
            "enum": [
              "cod",
              "bank_transfer",
              "wallet",
              "momo",
              "vnpay",
              "payos"
            ],
            "type": "string"
          },
          "transaction_code": {
            "type": "string"
          },
          "status": {
            "enum": [
              "pending",
              "success",
              "failed"
            ],
            "type": "string"
          },
          "type": {
            "enum": [
              "order_payment",
              "wallet_topup"
            ],
            "type": "string"
          },
          "note": {
            "type": "string"
          },
          "paid_at": {
            "format": "date-time",
            "type": "string"
          },
          "payos_order_code": {
            "type": "string"
          },
          "payos_payment_link_id": {
            "type": "string"
          },
          "payos_checkout_url": {
            "type": "string"
          },
          "payos_qr_code": {
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "order",
          "user",
          "amount",
          "payment_method",
          "transaction_code",
          "status",
          "type",
          "note",
          "paid_at",
          "payos_order_code",
          "payos_payment_link_id",
          "payos_checkout_url",
          "payos_qr_code",
          "created_at",
          "updated_at"
        ]
      },
      "UpdatePaymentDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "success",
              "failed"
            ]
          },
          "transaction_code": {
            "type": "string"
          }
        }
      },
...
    }
  }
}